+++ packages/eslint-plugin-gestalt/src/__fixtures__/only-valid-tokens/invalid/dangerouslySetInlineStyle-input.js
@@ -1,0 +1,13 @@
+import { Box } from 'gestalt';
+
+export default function TestElement() {
+  return (
+    <Box
+      dangerouslySetInlineStyle={{
+        __style: {
+          backgroundColor: 'var(--color-border-container)',
+        },
+      }}
+    />
+  );
+}