+++ packages/eslint-plugin-gestalt/src/__fixtures__/only-valid-tokens/invalid/unsafeCss-input.js
@@ -1,0 +1,11 @@
+import { Box } from 'gestalt';
+
+export default function TestElement() {
+  const getTableInlineStyle = () => `
+.CustomColor tbody tr td > button[aria-label="collapse"] {
+  background-color: var(--color-background-default);
+}
+`;
+
+  return <div />;
+}