+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 146.2.0 (Mar 11, 2024)
+
+### Minor
+
+- ESLint plugin: `only-valid-tokens` ([#3477](https://github.com/pinterest/gestalt/pull/3477)) - [Preview link](https://deploy-preview-3477--gestalt.netlify.app?devexample=true)
+
+## 146.1.1 (Mar 11, 2024)
+
+### Patch
+
+- Masonry: Enable two column module on secod batch ([#3473](https://github.com/pinterest/gestalt/pull/3473)) - [Preview link](https://deploy-preview-3473--gestalt.netlify.app?devexample=true)
+
 ## 146.1.0 (Mar 6, 2024)
 
 ### Minor